requirements.txt
pywin32
pystray
pillow